requirements.txt
pyserial>=3.5
pandas>=1.5.0
matplotlib>=3.6.0
numpy>=1.23.0